delete not used file, add Services class, and HomePage
