+++ lib/agenda/main.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'pages/agenda_pages_origin.dart';
-import 'pages/agenda_pages.dart';
-import 'pages/another_agenda_pages.dart';
-import 'pages/detail_agenda.dart';
+import 'pages/home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -20,7 +17,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home: const AgendaAinx(),
+      home: const HomePage(),
     );
   }
 }
