+++ lib/agenda/pages/home_page.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'agenda_page.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: [
+            ElevatedButton(
+              onPressed: () {
+                // Impelement to Tunjangan Page
+              },
+              child: const Text("Tunjangan"),
+            ),
+            ElevatedButton(
+              onPressed: () => Navigator.of(context).push(
+                MaterialPageRoute(builder: (context) => const AgendaPage()),
+              ),
+              child: const Text("Agenda"),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
